Add chart component Make flex width of the chart component correctly
